Fix Word class
* use BitSet
* byte is 6 bit long

+++ emulator/instruction/src/main/java/fr/awildelephant/mix/emulator/instruction/Modification.java
@@ -1,8 +1,9 @@
 package fr.awildelephant.mix.emulator.instruction;
 
-public record Modification(byte value) {
+public record Modification(int value) {
 
     public FieldSpecification toFieldSpecification() {
+        // TODO: use bitwise operators
         return new FieldSpecification(value / 8, value % 8);
     }
 }
